add fp64 for conv2d_transpose op

+++ python/paddle/fluid/tests/unittests/test_conv2d_transpose_op.py
@@ -119,8 +119,8 @@ class TestConv2dTransposeOp(OpTest):
         self.init_op_type()
         self.init_test_case()
 
-        input_ = np.random.random(self.input_size).astype("float32")
-        filter_ = np.random.random(self.filter_size).astype("float32")
+        input_ = np.random.random(self.input_size).astype(self.dtype)
+        filter_ = np.random.random(self.filter_size).astype(self.dtype)
 
         self.inputs = {'Input': input_, 'Filter': filter_}
         self.attrs = {
@@ -138,7 +138,7 @@ class TestConv2dTransposeOp(OpTest):
             self.attrs['output_size'] = self.output_size
 
         output = conv2dtranspose_forward_naive(input_, filter_,
-                                               self.attrs).astype('float32')
+                                               self.attrs).astype(self.dtype)
 
         self.outputs = {'Output': output}
